refactor: extract PDF region materialization
